Added ryanf.cpp that modifies jump need to work on switching to jump sprite
