+++ Makefile
@@ -12,7 +12,7 @@ LFLAGS = -lrt -lX11 -lGLU -lGL -pthread -lm #-lXrandr
 all: campusrun
 
 campusrun: campusrun.cpp
-	g++ $(CFLAGS) campusrun.cpp ppm.cpp log.cpp libggfonts.a -Wall -Wextra $(LFLAGS) -o campusrun
+	g++ $(CFLAGS) campusrun.cpp header.h ppm.cpp log.cpp ryanf.cpp libggfonts.a -Wall -Wextra $(LFLAGS) -o campusrun
 clean:
 	rm -f campusrun
 	rm -f *.o
